[#780] Move string operations into common and windows platform
